Minor fix resolve path before searching things
When searching in relative paths, it's possible to end up
with duplicate paths if they aren't resolved first.

In order to fix this, the path are resolved right after they
are computed from the current work dir.

As a result, found addons will all have a resolved path
including the addons_path. In the end, there won't be any duplicate.